+++ NebulaModel/Packets/Session/LobbyRequest.cs
@@ -1,12 +1,9 @@
 ﻿using NebulaAPI;
 using NebulaModel.Networking;
 
-/*
- * This packet is only here to handle older clients and tell them to upgrade to a newer nebula version. (this packet was replaced by the lobby packets)
- */
 namespace NebulaModel.Packets.Session
 {
-    public class HandshakeRequest
+    public class LobbyRequest
     {
         public string Username { get; set; }
         public Float4[] MechaColors { get; set; }
@@ -14,10 +11,8 @@ namespace NebulaModel.Packets.Session
         public int ModsCount { get; set; }
         public int GameVersionSig { get; set; }
         public byte[] ClientCert { get; set; }
-
-        public HandshakeRequest() { }
-
-        public HandshakeRequest(byte[] clientCert, string username, Float4[] mechaColors)
+        public LobbyRequest() { }
+        public LobbyRequest(byte[] clientCert, string username, Float4[] mechaColors)
         {
             Username = username;
             MechaColors = mechaColors;