+++ NebulaModel/Packets/Universe/DysonLaunchDataPacket.cs
@@ -1,5 +1,8 @@
-﻿namespace NebulaModel.Packets.Universe
+﻿using NebulaAPI;
+
+namespace NebulaModel.Packets.Universe
 {
+    [HidePacketInDebugLogs]
     public class DysonLaunchDataPacket
     {
         public int Count { get; set; }