+++ NebulaNetwork/Server.cs
@@ -4,6 +4,7 @@ using NebulaModel.DataStructures;
 using NebulaModel.Networking;
 using NebulaModel.Packets.GameHistory;
 using NebulaModel.Packets.GameStates;
+using NebulaModel.Packets.Universe;
 using NebulaModel.Utils;
 using NebulaWorld;
 using System.Collections.Generic;
@@ -20,11 +21,13 @@ namespace NebulaNetwork
         private const float GAME_RESEARCH_UPDATE_INTERVAL = 2;
         private const float STATISTICS_UPDATE_INTERVAL = 1;
         private const float LAUNCH_UPDATE_INTERVAL = 2;
+        private const float DYSONSPHERE_UPDATE_INTERVAL = 5;
 
         private float gameStateUpdateTimer = 0;
         private float gameResearchHashUpdateTimer = 0;
         private float productionStatisticsUpdateTimer = 0;
-        private float dysonLaunchUpateTimer = 0;
+        private float dysonLaunchUpateTimer = 1;
+        private float dysonSphereUpdateTimer = 0;
 
         private readonly int port;
         private readonly bool loadSaveFile;
@@ -92,7 +95,7 @@ namespace NebulaNetwork
                     break;
 
                 case ConnectionState.Connecting:
-                    if (Multiplayer.Session.IsGameLoaded == false)
+                    if (Multiplayer.Session.IsGameLoaded == false && Multiplayer.Session.IsInLobby == false)
                     {
                         Sockets.CloseConnection(info.connection, (int)DisconnectionReason.HostStillLoading, "Host still loading, please try again later.", true);
                     }
@@ -177,6 +180,7 @@ namespace NebulaNetwork
             gameResearchHashUpdateTimer += Time.deltaTime;
             productionStatisticsUpdateTimer += Time.deltaTime;
             dysonLaunchUpateTimer += Time.deltaTime;
+            dysonSphereUpdateTimer += Time.deltaTime;            
 
             if (gameStateUpdateTimer > GAME_STATE_UPDATE_INTERVAL)
             {
@@ -204,6 +208,20 @@ namespace NebulaNetwork
             {
                 dysonLaunchUpateTimer = 0;
                 Multiplayer.Session.Launch.SendBroadcastIfNeeded();
+            }
+
+            if (dysonSphereUpdateTimer > DYSONSPHERE_UPDATE_INTERVAL)
+            {
+                dysonSphereUpdateTimer = 0;
+                DysonSphere[] dysonSpheres = GameMain.data.dysonSpheres;
+                for (int i = 0; i < dysonSpheres.Length; i++)
+                {
+                    DysonSphere dysonSphere = dysonSpheres[i];
+                    if (dysonSphere != null && (dysonSphere.energyReqCurrentTick + dysonSphere.energyGenCurrentTick > 0))
+                    {
+                        SendPacketToStar(new DysonSphereStatusPacket(dysonSphere), dysonSphere.starData.id);
+                    }
+                }
             }
 
             lock (Sockets)